Exact symbolic growth, unified reduction size contracts, and deterministic solver backends (#1137)
* Unify symbolic reduction, registry, and solver contracts

Introduce exact symbolic expressions, rule-owned size relations, strict extraction and numeric contracts, model-owned construction metadata, and deterministic solver selection as one compile-time contract. Migrate directly from the main-branch APIs without recording discarded intermediate designs.

* Expose registry-driven CLI and MCP workflows

Use the unified registry for creation, inspection, deterministic solver selection, reduction paths, and size reporting. Keep CLI and MCP outputs consistent and cover their user-facing error paths and round trips.

* Refresh end-to-end reduction examples

Update executable examples and integration scenarios for the final size, extraction, and registry contracts, including source-relative path assertions and exact variant handling.

* Document the unified reduction workflows

Align design, CLI, MCP, paper, and maintainer workflow documentation with exact expressions, rule-owned size relations, registry-owned construction, and deterministic solver dispatch.

* Harden cross-platform build verification

Use the required HiGHS backend, keep benchmark-only dependencies out of ordinary builds, add portable stack handling, and exercise macOS ARM64, Windows x86_64, and RISC-V targets in CI.

* Restore customized solver naming

* Refine path selection and solver outcomes

* Track symbolic growth precision and remove unsound rule

* Simplify symbolic Big-O comparison

* Propagate polynomial size bounds through subtraction

* Limit path output requests to 999

* Bound Pareto path search

* Clarify registered solver dispatch

* Validate rounded ILP solutions

* Complete symbolic reduction size contracts

* refactor: enforce numeric and error domain contracts

* refactor: define typed problem and ILP boundaries

* fix: address PR #1137 CI failure

Use the Rust 1.98 slice chunk API required by Clippy and remove the numeric-format sections from the problem and rule issue templates.

* refactor: replace size contracts with problem parameters

* fix: standardize factoring and closest vector inputs

Derive factoring widths from the target and canonicalize factor witnesses. Model CVP without artificial coefficient bounds and register exact sphere enumeration. Make QUBO numeric variants and same-problem variant reductions explicit so the standard models remain executable.

* fix: preserve feasibility across ILP reductions

Disambiguate infeasible and unbounded ILPs with a zero-objective solve. Map optimal target values through aggregate reductions before extracting decision witnesses, including HamiltonianCircuit and PartitionIntoCliques. Cover decision bounds and error propagation with regression tests.

* fix: align ILP models and solver pipelines

* Improve exact solver coverage for larger instances

* fix: correct validated reductions and register ensemble ILP solver

Align model validation, flow and scheduling reductions with their mathematical contracts. Add the EnsembleComputation ILP rule and fixed solver pipeline, update proofs and regression tests, and propagate construction overflow through shared error APIs.

* Fix concrete variant registration, CLI ordering, and dominating-set reduction

* Fix reduction mappings and simplify rule implementations

Repair reduction constructions, witness extraction, and boundary handling identified during corpus validation. Include the associated model and variant support, exact JSON replay, regression tests, and mathematical documentation.

Reuse existing SAT allocation and big-integer APIs, remove redundant extraction state, and avoid allocating zero-coefficient sets.

Validation: make check (6555 tests passed), make coverage, and git diff --cached --check. Known unresolved rules and resource limits remain tracked separately.

* Use array chunks for ensemble solution extraction

* Typeset website equations and registry expressions with MathML

* Fix solver edge cases and concrete variant construction

Keep ILP numerical semantics explicit, report unresolved decisions, check circulation overflow, and enumerate nearby CVP candidates first.

Make unit weights implicit across One variants and share concrete construction inputs across CLI, MCP, and decision wrappers. Add regression coverage and document the contracts.

* Expose instance parameter values in pred inspect

* fix: address eight correctness findings in PR #1137

---------

+++ src/rules/unitdiskmapping/grid.rs
@@ -9,13 +9,13 @@ pub enum CellState {
     #[default]
     Empty,
     Occupied {
-        weight: i32,
+        weight: i64,
     },
     Doubled {
-        weight: i32,
+        weight: i64,
     },
     Connected {
-        weight: i32,
+        weight: i64,
     },
 }
 
@@ -28,7 +28,7 @@ impl CellState {
         !self.is_empty()
     }
 
-    pub fn weight(&self) -> i32 {
+    pub fn weight(&self) -> i64 {
         match self {
             CellState::Empty => 0,
             CellState::Occupied { weight } => *weight,
@@ -117,7 +117,7 @@ impl MappingGrid {
     /// For unweighted mode, all weights are 1 so this doesn't matter.
     ///
     /// Silently ignores out-of-bounds access.
-    pub fn add_node(&mut self, row: usize, col: usize, weight: i32) {
+    pub fn add_node(&mut self, row: usize, col: usize, weight: i64) {
         if row < self.rows && col < self.cols {
             match self.content[row][col] {
                 CellState::Empty => {
